% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCE.r
\name{PROCE}
\alias{PROCE}
\title{Ruta de 02. Procedimientos}
\usage{
PROCE()
}
\arguments{
\item{Ninguno}{No se requieren parametros}
}
\value{
Ruta de carpeta 02. Procedimientos
}
\description{
Ruta de 02. Procedimientos
}
