% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODU.r
\name{PRODU}
\alias{PRODU}
\title{Ruta de 03. Productos}
\usage{
PRODU(Productos)
}
\arguments{
\item{Ninguno}{No se requieren parametros}
}
\value{
Ruta de carpeta 03. Productos
}
\description{
Ruta de 03. Productos
}
