% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruta_dinamica_list.r
\name{ruta_dinamica_list}
\alias{ruta_dinamica_list}
\title{Ruta de carpetas}
\usage{
ruta_dinamica_list()
}
\arguments{
\item{Ninguno}{No se requieren parametros}
}
\value{
Ruta de carpetas en una lista
}
\description{
Ruta de carpetas
}
